+++ sdk/communication/communication-identity/src/common/tracing.ts
@@ -1,13 +1,10 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT license.
 
-import { createSpanFunction } from "@azure/core-tracing";
+import { createTracingClient } from "@azure/core-tracing";
 
-/**
- * Creates a span using the global tracer.
- * @internal
- */
-export const createSpan = createSpanFunction({
-  packagePrefix: "Azure.Communication",
-  namespace: "Microsoft.Communication",
+export const tracingClient = createTracingClient({
+  namespace: "Azure.Communication",
+  packageName: "@azure/communication-identity",
+  packageVersion: "1.1.2"
 });