+++ sdk/communication/communication-identity/src/communicationIdentityClient.ts
@@ -17,9 +17,8 @@ import {
 import { InternalClientPipelineOptions, OperationOptions } from "@azure/core-client";
 import { KeyCredential, TokenCredential, isTokenCredential } from "@azure/core-auth";
 import { IdentityRestClient } from "./generated/src/identityRestClient";
-import { SpanStatusCode } from "@azure/core-tracing";
-import { createSpan } from "./common/tracing";
 import { logger } from "./common/logger";
+import { tracingClient } from "./common/tracing";
 
 const isCommunicationIdentityClientOptions = (
   options: any
@@ -101,27 +100,18 @@ export class CommunicationIdentityClient {
    * @param scopes - Scopes to include in the token.
    * @param options - Additional options for the request.
    */
-  public async getToken(
+  public getToken(
     user: CommunicationUserIdentifier,
     scopes: TokenScope[],
     options: OperationOptions = {}
   ): Promise<CommunicationAccessToken> {
-    const { span, updatedOptions } = createSpan("CommunicationIdentity-issueToken", options);
-    try {
+    return tracingClient.withSpan("CommunicationIdentity-issueToken", options, async (updatedOptions) => {
       return await this.client.communicationIdentityOperations.issueAccessToken(
         user.communicationUserId,
         scopes,
         updatedOptions
       );
-    } catch (e: any) {
-      span.setStatus({
-        code: SpanStatusCode.ERROR,
-        message: e.message,
-      });
-      throw e;
-    } finally {
-      span.end();
-    }
+    });
   }
 
   /**
@@ -130,25 +120,16 @@ export class CommunicationIdentityClient {
    * @param user - The user whose tokens are being revoked.
    * @param options - Additional options for the request.
    */
-  public async revokeTokens(
+  public revokeTokens(
     user: CommunicationUserIdentifier,
     options: OperationOptions = {}
   ): Promise<void> {
-    const { span, updatedOptions } = createSpan("CommunicationIdentity-revokeTokens", options);
-    try {
+    return tracingClient.withSpan("CommunicationIdentity-revokeTokens", options, async (updatedOptions) => {
       await this.client.communicationIdentityOperations.revokeAccessTokens(
         user.communicationUserId,
         updatedOptions
       );
-    } catch (e: any) {
-      span.setStatus({
-        code: SpanStatusCode.ERROR,
-        message: e.message,
-      });
-      throw e;
-    } finally {
-      span.end();
-    }
+    });
   }
 
   /**
@@ -156,22 +137,13 @@ export class CommunicationIdentityClient {
    *
    * @param options - Additional options for the request.
    */
-  public async createUser(options: OperationOptions = {}): Promise<CommunicationUserIdentifier> {
-    const { span, updatedOptions } = createSpan("CommunicationIdentity-createUser", options);
-    try {
+  public createUser(options: OperationOptions = {}): Promise<CommunicationUserIdentifier> {
+    return tracingClient.withSpan("CommunicationIdentity-createUser", options, async (updatedOptions) => {
       const result = await this.client.communicationIdentityOperations.create(updatedOptions);
       return {
         communicationUserId: result.identity.id,
       };
-    } catch (e: any) {
-      span.setStatus({
-        code: SpanStatusCode.ERROR,
-        message: e.message,
-      });
-      throw e;
-    } finally {
-      span.end();
-    }
+    });
   }
 
   /**
@@ -184,11 +156,8 @@ export class CommunicationIdentityClient {
     scopes: TokenScope[],
     options: OperationOptions = {}
   ): Promise<CommunicationUserToken> {
-    const { span, updatedOptions } = createSpan(
-      "CommunicationIdentity-createUserAndToken",
-      options
-    );
-    try {
+    return tracingClient.withSpan(
+      "CommunicationIdentity-createUserAndToken", options, async (updatedOptions) => {
       const { identity, accessToken } = await this.client.communicationIdentityOperations.create({
         createTokenWithScopes: scopes,
         ...updatedOptions,
@@ -197,15 +166,7 @@ export class CommunicationIdentityClient {
         ...accessToken!,
         user: { communicationUserId: identity.id },
       };
-    } catch (e: any) {
-      span.setStatus({
-        code: SpanStatusCode.ERROR,
-        message: e.message,
-      });
-      throw e;
-    } finally {
-      span.end();
-    }
+    });
   }
 
   /**
@@ -214,25 +175,16 @@ export class CommunicationIdentityClient {
    * @param user - The user being deleted.
    * @param options - Additional options for the request.
    */
-  public async deleteUser(
+  public deleteUser(
     user: CommunicationUserIdentifier,
     options: OperationOptions = {}
   ): Promise<void> {
-    const { span, updatedOptions } = createSpan("CommunicationIdentity-deleteUser", options);
-    try {
+    return tracingClient.withSpan("CommunicationIdentity-deleteUser", options, async (updatedOptions) => {
       await this.client.communicationIdentityOperations.delete(
         user.communicationUserId,
         updatedOptions
       );
-    } catch (e: any) {
-      span.setStatus({
-        code: SpanStatusCode.ERROR,
-        message: e.message,
-      });
-      throw e;
-    } finally {
-      span.end();
-    }
+    });
   }
 
   /**
@@ -243,26 +195,14 @@ export class CommunicationIdentityClient {
   public async getTokenForTeamsUser(
     options: GetTokenForTeamsUserOptions
   ): Promise<CommunicationAccessToken> {
-    const { span, updatedOptions } = createSpan(
-      "CommunicationIdentity-getTokenForTeamsUser",
-      options
-    );
-    const { teamsUserAadToken, clientId, userObjectId } = options;
-    try {
+    return tracingClient.withSpan("CommunicationIdentity-getTokenForTeamsUser", options, async (updatedOptions) => {
+      const { teamsUserAadToken, clientId, userObjectId } = updatedOptions;
       return await this.client.communicationIdentityOperations.exchangeTeamsUserAccessToken(
         teamsUserAadToken,
         clientId,
         userObjectId,
         updatedOptions
       );
-    } catch (e: any) {
-      span.setStatus({
-        code: SpanStatusCode.ERROR,
-        message: e.message,
-      });
-      throw e;
-    } finally {
-      span.end();
-    }
+    });
   }
 }