+++ sdk/communication/communication-common/CHANGELOG.md
@@ -1,5 +1,17 @@
 # Release History
 
+## 2.1.2 (Unreleased)
+
+- Updated to `@azure/core-tracing` 1.0.
+
+### Features Added
+
+### Breaking Changes
+
+### Bugs Fixed
+
+### Other Changes
+
 ## 2.1.1 (Unreleased)
 
 ### Features Added