+++ sdk/communication/communication-identity/CHANGELOG.md
@@ -1,5 +1,17 @@
 # Release History
 
+## 1.1.2 (Unreleased)
+
+### Features Added
+
+- Updated to `@azure/core-tracing` 1.0.
+
+### Breaking Changes
+
+### Bugs Fixed
+
+### Other Changes
+
 ## 1.1.1 (Unreleased)
 
 ### Features Added